Fix build on OSX Lion * fix conflicting typdef error in ACE * fix ambiguous ctor reference error * prevent assert when changing thread priority is not supported

+++ src/game/MapPersistentStateMgr.cpp
@@ -779,7 +779,7 @@ void MapPersistentStateManager::PackInstances()
         delete result;
     }
 
-    BarGoLink bar(InstanceSet.size() + 1);
+    BarGoLink bar(uint64(InstanceSet.size() + 1));
     bar.step();
 
     uint32 InstanceNumber = 1;
